Cleanup and refactor.
