+++ lib/rutabaga/example_group/feature.rb
