+++ lib/rutabaga/util.rb
@@ -35,3 +35,9 @@ module Rutabaga
     end
   end
 end
+
+::RSpec.configure do |c|
+  c.include Rutabaga::Feature
+  # Blow away turnip's pattern, and focus just on features directory
+  c.pattern.gsub!(",**/*.feature", ",features/**/*.feature")
+end
